+++ src/utils/interfaces.ts
@@ -127,7 +127,7 @@ export const Pets = [
         user_id: 1,
         Id: 452568,
         name: 'Alfred',
-        pet_type:'Dog',
+        pet_type:'dog',
         breed: 'Golden Retriever',
         birthday: '05/05/2015',
         symptoms: ['Itching', 'Hair loss', 'Redness'],
@@ -142,7 +142,7 @@ export const Pets = [
         user_id: 1,
         Id: 4586,
         name: 'Bella',
-        pet_type:'Cat',
+        pet_type:'cat',
         breed: 'Siamese',
         birthday: '05/05/2015',
         symptoms: ['Itching', 'Hair loss', 'Redness'],
@@ -157,7 +157,7 @@ export const Pets = [
         user_id: 1,
         Id: 1285,
         name: 'Charlie',
-        pet_type:'Dog',
+        pet_type:'dog',
         breed: 'Thoroughbred',
         birthday: '05/05/2015',
         symptoms: ['Itching', 'Hair loss', 'Redness'],
